accountbl finished without testing

+++ MSPSSERPnju/MSPSSCommonModule/src/main/java/po/AccountPO.java
@@ -11,7 +11,7 @@ import java.util.ArrayList;
 public class AccountPO implements Serializable {
     private int id;
     private String name;
-    private int money;
+    private double money;
     private String createtime;
 
     public AccountPO(){
@@ -19,7 +19,7 @@ public class AccountPO implements Serializable {
     }
 
 
-    public AccountPO(String name,int money,String createtime){
+    public AccountPO(String name,double money,String createtime){
         this.name = name;
         this.money = money;
 
@@ -42,11 +42,11 @@ public class AccountPO implements Serializable {
         this.name = name;
     }
 
-    public int getMoney() {
+    public double getMoney() {
         return money;
     }
 
-    public void setMoney(int money) {
+    public void setMoney(double money) {
         this.money = money;
     }
